feat: split state functions into multiple files
